[PR Feedback] rename source file client to webhook_client
