+++ components/builder-notify/src/lib.rs
@@ -5,17 +5,17 @@ extern crate log;
 extern crate serde_derive;
 
 pub mod cli;
-pub mod client;
 pub mod config;
 pub mod error;
 pub mod hook;
 pub mod hub;
 pub mod server;
+pub mod webhook_client;
 
-use crate::{client::WebhookClient,
-            config::Config,
+use crate::{config::Config,
             hook::Webhook,
-            hub::Hub};
+            hub::Hub,
+            webhook_client::WebhookClient};
 
 pub fn get_hub(config: &Config) -> Hub {
     debug!("NotifyConfig {:?}", config);
